AMP-15175
add null guards when tearing down FastFilterCacher

git-svn-id: file:///svn/branches/AMP_2_4_RELEASE@20328 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/dgfoundation/amp/ar/filtercacher/FastFilterCacher.java
@@ -1,5 +1,6 @@
 package org.dgfoundation.amp.ar.filtercacher;
 
+import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
 
@@ -54,7 +55,11 @@ public class FastFilterCacher extends FilterCacher {
 			return; //nothing to delete
 		try
 		{
-			Statement statement = getConnection().createStatement();
+			Connection conn = getConnection();
+			if (conn == null)
+				return; //shouldn't happen, but let's guard against it
+			
+			Statement statement = conn.createStatement();
 			String stat = "DROP TABLE IF EXISTS" + tempTableName;
 			statement.executeQuery(stat);
 			statement.close();
